pyproject.toml
[project]
name = "backgemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.32.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "streamlit>=1.49.1",
]
